Add StrWithHeader counterpart to SliceWithHeader
